requirements.txt
numpy==1.24.3
altair==4.2.2
TTS==0.22.0
openai-whisper==20240930
moviepy==1.0.3
transformers==4.47.0
streamlit==1.18.0
validators==0.34.0
ffmpeg==1.4
nltk==3.9.1